+++ noodles-tabix/src/io.rs
@@ -1,0 +1,3 @@
+//! Tabix I/O.
+
+pub mod indexed_reader;