+++ noodles-tabix/src/io/indexed_reader.rs
@@ -1,0 +1,5 @@
+//! Tabix indexed reader.
+
+mod builder;
+
+pub use self::builder::Builder;