+++ noodles-tabix/src/lib.rs
@@ -29,6 +29,7 @@
 pub mod r#async;
 
 pub mod index;
+pub mod io;
 mod reader;
 mod writer;
 
@@ -37,7 +38,7 @@ pub use self::{reader::Reader, writer::Writer};
 #[cfg(feature = "async")]
 pub use self::r#async::{Reader as AsyncReader, Writer as AsyncWriter};
 
-use std::{fs::File, io, path::Path};
+use std::{fs::File, path::Path};
 
 use noodles_csi::Index;
 
@@ -56,7 +57,7 @@ static MAGIC_NUMBER: &[u8] = b"TBI\x01";
 /// let index = tabix::read("sample.vcf.gz.tbi")?;
 /// # Ok::<(), io::Error>(())
 /// ```
-pub fn read<P>(src: P) -> io::Result<Index>
+pub fn read<P>(src: P) -> std::io::Result<Index>
 where
     P: AsRef<Path>,
 {
@@ -83,7 +84,7 @@ where
 /// tabix::write("sample.vcf.gz.tbi", &index)?;
 /// # Ok::<(), io::Error>(())
 /// ```
-pub fn write<P>(dst: P, index: &Index) -> io::Result<()>
+pub fn write<P>(dst: P, index: &Index) -> std::io::Result<()>
 where
     P: AsRef<Path>,
 {